requirements.txt
adversarial_robustness_toolbox==1.19.0
art==6.4
lime==0.2.0.1
matplotlib==3.10.0
numpy==2.2.1
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
skimage==0.0
torch==2.2.2
torchvision==0.17.2
